reorganize tests, add webtest for rest api tests
