+++ docker/dockerfiles/Dockerfile.server
@@ -56,9 +56,7 @@ COPY mkdocs.yml mkdocs.yml
 COPY monitor.py monitor.py
 COPY scripts scripts
 COPY setup.py setup.py
-COPY test_cli.py test_cli.py
 COPY test_runner.py test_runner.py
-COPY stress_test.py stress_test.py
 COPY tests tests
 COPY views views
