+++ requirements-server.txt
@@ -13,6 +13,7 @@ boto3==1.14.35
 # For testing
 mock==4.0.2
 nose==1.3.7
+webtest==2.0.35
 
 # Server-specific
 alembic==1.4.2
