+++ tests/unit/client/json_api_client_test.py
