+++ tests/unit/files/a-symlink.txt
