+++ tests/unit/files/dir1/broken-symlink
