+++ tests/unit/files/dir2/the-only-file
