+++ tests/unit/files/ignore_test/__MACOSX/ignored.txt
