+++ tests/unit/files/ignore_test/dir/__MACOSX/ignored.txt
