+++ tests/unit/files/ignore_test/dir/ignored2.txt
