+++ tests/unit/files/ignore_test/dir/not_ignored2.txt
