+++ tests/unit/files/ignore_test/ignored.txt
