+++ tests/unit/files/ignore_test/ignored_dir/numbers.txt
