+++ tests/unit/files/ignore_test/not_ignored.txt
