+++ tests/unit/files/netcat-test.py
