+++ tests/unit/lib/bundle_cli_test.py
