+++ tests/unit/lib/canonicalize_test.py
