+++ tests/unit/lib/cli_util_test.py
