+++ tests/unit/lib/interactive_session_test.py
