+++ tests/unit/lib/upload_manager_test.py
