+++ tests/unit/objects/dependency_test.py
