+++ tests/unit/rest/util_test.py
