+++ tests/unit/server/bundle_manager_test.py
